backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
passlib==1.7.4
bcrypt==3.2.2
python-jose[cryptography]
python-dotenv
requests
